Pulling common logic from RocksDBLookupExtractor to base
